requirements.txt
# Основные зависимости
python-telegram-bot==20.7
psycopg2-binary==2.9.9
aiohttp==3.9.3
python-dotenv==1.0.1
python-dateutil==2.9.0

# Дополнительные утилиты
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17

# Для работы с датами
pytz==2023.3.post1

# Для разработки и тестирования
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Безопасность
cryptography==41.0.7

# Логирование
structlog==23.2.0

# Для работы с JSON
ujson==5.8.0

# Версия Python
python_version >= "3.11"